Test Factory.PrintObject by fake command.
