+++ hack/verify-flags/known-flags.txt
@@ -467,6 +467,7 @@ show-all
 show-events
 show-kind
 show-labels
+show-namespace
 shutdown-fd
 shutdown-fifo
 since-seconds
